+++ test/support/factory.ex
@@ -10,6 +10,7 @@ defmodule TdDd.Factory do
   alias TdCx.Events.Event
   alias TdCx.Jobs.Job
   alias TdCx.Sources.Source
+  alias TdDd.DataStructures.CatalogViewConfig
   alias TdDd.DataStructures.DataStructure
   alias TdDd.DataStructures.DataStructureLink
   alias TdDd.DataStructures.DataStructureRelation
@@ -781,6 +782,14 @@ defmodule TdDd.Factory do
     %TdDd.Access{
       source_user_name: sequence("access_source_user_name"),
       details: %{}
+    }
+    |> merge_attributes(attrs)
+  end
+
+  def catalog_view_config_factory(attrs) do
+    %CatalogViewConfig{
+      field_type: "metadata",
+      field_name: sequence("field_name")
     }
     |> merge_attributes(attrs)
   end