+++ test/td_dd/data_structures/search/aggregations_test.exs
@@ -3,6 +3,8 @@ defmodule TdDd.DataStructures.Search.AggregationsTest do
 
   alias TdDd.DataStructures.Search.Aggregations
 
+  @missing_term_name Aggregations.missing_term_name()
+
   @static_fields [
     "class.raw",
     "confidential.raw",
@@ -79,6 +81,22 @@ defmodule TdDd.DataStructures.Search.AggregationsTest do
         },
         ["metadata.foo", "metadata.bar", "metadata.baz"]
       )
+    end
+
+    test "includes custom catalog view configs field aggregations" do
+      insert(:catalog_view_config, field_type: "metadata", field_name: "database")
+      insert(:catalog_view_config, field_type: "mutable_metadata", field_name: "command")
+      insert(:catalog_view_config, field_type: "note", field_name: "layer")
+
+      %{
+        "metadata.database" => %{
+          terms: %{field: "metadata.database.keyword", missing: @missing_term_name}
+        },
+        "mutable_metadata.command" => %{
+          terms: %{field: "mutable_metadata.command.keyword", missing: @missing_term_name}
+        },
+        "note.layer" => %{terms: %{field: "note.layer.raw", missing: @missing_term_name}}
+      } = Aggregations.aggregations()
     end
   end
 end