+++ test/td_dq_web/controllers/rule_filter_controller_test.exs
@@ -30,7 +30,7 @@ defmodule TdDqWeb.RuleFilterControllerTest do
                |> post(Routes.rule_filter_path(conn, :search, %{"filters" => filters}))
                |> json_response(:ok)
 
-      assert data == %{"domain_id" => %{"values" => [1, 2]}}
+      assert %{"domain_id" => %{"values" => [1, 2]}} = data
     end
 
     @tag authentication: [role: "user", permissions: ["view_quality_rule"]]